Solution to Calculate Employee Bonus

+++ Calculate Employee Bonus@@67877d61dc107486ad5c274f/code.c
@@ -2,7 +2,7 @@
 struct Employee{
     int id;
     char name[100];
-    double salary;
+    int salary;
     float bonus;
 };
 int main(){
